Created output File and added code to write to the file
Each line of code I added has its comment as to what it does. I tried creating a class and to implement the class, but I was having difficulties with that, so I just directly write to the file whenever I need to. Maybe changed to a class (that works) for next deliverable, depending on how easier it makes the readability + runtime.