few modifications of the code
in alphabeta, the cloning of the game isntance is done once at the beginning of the method and both minimizing and maximizing use the same reference to the temp variable, also made a "allowed time" variable to make readability a bit more fluid. fixed the issue with minimizing in alphabeta, troubleshot some other issue. this current version allows a playthrough of the attacker mode, but it ends at "the computer doesn't know what to do" after we input the first move.